numa-device-plugin: move to production paths
Move NUMACell device plugin from test/deviceplugin/ to production
paths. numacell/main.go becomes the binary entry point with
numacell/pkg/ subtree for packages (dpm, plugin, cmd).
Shared packages (api, manifests) placed in pkg/numacell/.

Updates to  all consumer import paths.
